requirements.txt
APScheduler==3.6.3
asgiref==3.2.7
astroid==2.4.1
autopep8==1.5.2
beautifulsoup4==4.9.1
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
defusedxml==0.6.0
Django==3.0.6
django-allauth==0.41.0
django-crispy-forms==1.9.1
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
msal==1.3.0
oauthlib==3.1.0
pycparser==2.20
PyJWT==1.7.1
pycodestyle==2.6.0
pycparser==2.20
PyJWT==1.7.1
pylint==2.5.2
python3-openid==3.1.0
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
soupsieve==2.0.1
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
tzlocal==2.1
urllib3==1.25.9
wrapt==1.12.1
bs4==0.0.1
apscheduler==3.6.3